move predict tools into separate files
add notebook for muon reco
