+++ pyproject.toml
@@ -62,8 +62,8 @@ documentation = "https://ctlearn.readthedocs.io/en/latest/"
 
 [project.scripts]
 ctlearn-train-model = "ctlearn.tools.train_model:main"
-ctlearn-predict-mono-model = "ctlearn.tools.predict_model:mono_tool"
-ctlearn-predict-stereo-model = "ctlearn.tools.predict_model:stereo_tool"
+ctlearn-predict-mono-model = "ctlearn.tools.predict_mono_model:main"
+ctlearn-predict-stereo-model = "ctlearn.tools.predict_stereo_model:main"
 ctlearn-predict-LST1= "ctlearn.tools.predict_LST1:main"
 
 [tool.setuptools_scm]
